refactor(mcp): split tool_context.py into a package + fix truncation O(n^2)
Break the 1213-line get_context module into an orchestrator + focused
sibling modules behind a package façade. `from ...tool_context import
get_context` resolves unchanged; the truncation helpers tests import
(_truncate_to_budget, _CHAR_BUDGET) stay importable from the package root.

- truncation.py: budget constants + _truncate_to_budget (+ symbol-priority
  helpers). PERF FIX: stage 2 no longer re-serialises the whole response per
  candidate symbol (O(targets x symbols^2)). It precomputes each symbol's
  compact-JSON cost once and tracks a running sum — the whole-response size
  with a kept set is exactly base + sum(costs) + (len-1), so the keep/drop
  decision is byte-identical to the old _size()-per-symbol loop.
- targets.py: _resolve_one_target + _synthesize_structural_summary.
- enrichment.py: _resolve_call_graph / _metrics / _community / _health.
- kg.py: layer / tour / role helpers.
- context.py: get_context orchestrator + cross-repo enrichment.

Verbatim moves otherwise. New test embeds the original O(n^2) implementation
as an oracle and asserts byte-identical output across 6 over-budget fixtures.
tests/unit/server (265) green; ruff clean on new modules; no net-new mypy
errors (page union-attr notes in targets.py are verbatim-moved from the old
module branch; the 2 in mcp_server/__init__.py are pre-existing).
